+++ extensions/podman/packages/extension/src/helpers/podman-binary-location-helper.ts
