+++ extensions/podman/packages/extension/src/helpers/podman-info-helper.ts
