+++ extensions/podman/packages/extension/src/helpers/wsl-helper.ts
@@ -18,7 +18,7 @@
 
 import * as extensionApi from '@podman-desktop/api';
 
-import { normalizeWSLOutput } from './util';
+import { normalizeWSLOutput } from '../util';
 
 export interface WSLVersionInfo {
   wslVersion?: string;
